Move vector space to its own file to try and avoid circular imports in base.py
