+++ geomstats/geometry/matrices.py
@@ -5,8 +5,8 @@ from functools import reduce
 import geomstats.backend as gs
 import geomstats.errors
 from geomstats.algebra_utils import flip_determinant, from_vector_to_diagonal_matrix
-from geomstats.geometry.base import VectorSpace
 from geomstats.geometry.euclidean import EuclideanMetric
+from geomstats.geometry.vector_space import VectorSpace
 
 
 class Matrices(VectorSpace):
